# Do not pre-allocate error vector in TRY